Working Layout and static previews
